chore: nightly e2es for ib dockerfiles (#1908)
## Description

We had a gap in the e2e testability in the Iron Bank images. This is
caused by Iron Bank expecting a release artifact that is pulled from
GitHub. This PR creates jobs that run nightly and perform e2e tests on
the IB Dockerfiles with the code from the main branch.

1. Pulls the Iron Bank Dockerfiles into Pepr
2. Builds the images
3. Runs the excellent examples with Pepr main and the Iron Bank
controller images.


## Related Issue

Fixes #1896 
<!-- or -->
Relates to #

## Type of change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [x] Other (security config, docs update, etc)

## Checklist before merging
- [x] Unit,
[Journey](https://github.com/defenseunicorns/pepr/tree/main/journey),
[E2E Tests](https://github.com/defenseunicorns/pepr-excellent-examples),
[docs](https://github.com/defenseunicorns/pepr/tree/main/docs),
[adr](https://github.com/defenseunicorns/pepr/tree/main/adr) added or
updated as needed
- [x] [Contributor Guide
Steps](https://docs.pepr.dev/main/contribute/#submitting-a-pull-request)
followed

---------

Signed-off-by: Case Wylie <cmwylie19@defenseunicorns.com>