chore: consolidate prettier config to one file for consistent behavior (#1952)
## Description

We had two different `.prettierrc` files which overrode each other on
the `bracketSameLine` rule that was in `src/lib/.prettierrc`. This PR
uses one `.prettierrc` file for the entire project and applies a
consistent ruleset.

It appears that `src/lib/` preferred the use of `bracketSameLine` but
the rule was not being applied due to `bracketSameLine` being `false` in
the top-level config. We may as well be consistent across the entire
project, so this PR enables that rule.

Note that this also required changes to the
`src/templates/prettierrc.json` template that new pepr modules use. This
is not expected to create any breaking changes to existing users of Pepr
since the template is used at module creation; new pepr modules will use
a different `.prettierrc` file. It may be worth noting this change to
users, tbd.

## Related Issue

Fixes #1900 

## Type of change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [x] Other (security config, docs update, etc)

## Checklist before merging
- [x] Unit,
[Journey](https://github.com/defenseunicorns/pepr/tree/main/journey),
[E2E Tests](https://github.com/defenseunicorns/pepr-excellent-examples),
[docs](https://github.com/defenseunicorns/pepr/tree/main/docs),
[adr](https://github.com/defenseunicorns/pepr/tree/main/adr) added or
updated as needed
- [x] [Contributor Guide
Steps](https://docs.pepr.dev/main/contribute/#submitting-a-pull-request)
followed