chore: refactor watch-processor tests for clarity (#2013)
## Description

This PR presented itself as a refactor while working on #1800. It should
be merged in _after_ that issue is resolved.

This PR was motivated from seeing hard to read test output in the
console when running unit tests. The old format was
[old.log](https://github.com/user-attachments/files/19608363/old.log)
and the new format is
[new-tests.log](https://github.com/user-attachments/files/19608374/new-tests.log).
This is possible due to a custom `toString()` function that was added to
`LoggableKubernetesObject`. Our source code expects a
`KubernetesObject`, and this extension of the class adheres to the same
specificiation.

## Related Issue

Relates to #1800 and #2014

## Type of change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [x] Other (security config, docs update, etc)

## Checklist before merging
- [x] Unit,
[Journey](https://github.com/defenseunicorns/pepr/tree/main/journey),
[E2E Tests](https://github.com/defenseunicorns/pepr-excellent-examples),
[docs](https://github.com/defenseunicorns/pepr/tree/main/docs),
[adr](https://github.com/defenseunicorns/pepr/tree/main/adr) added or
updated as needed
- [x] [Contributor Guide
Steps](https://docs.pepr.dev/main/contribute/#submitting-a-pull-request)
followed