chore: make uds test use library (#2024)
## Description

Pepr is testing using the library candidate in the integration tests,
however, it is not as sensitive to missing types as UDS Core.

As seen in this
[comment](https://github.com/defenseunicorns/pepr/issues/1989#issuecomment-2787472053)
a basic Pepr module will build if a type is excluded from the library,
but UDS Core will fail the build.

We already have a UDS test, where we install `slim-dev`, however, the
test was only testing the pepr release candidate controller image, and
not installing the release candidate library.

This PR results in UDS Core `slim-dev` package being build with the
release candidate

```json
  "dependencies": {
    "pepr": "file:pepr-0.0.0-development.tgz"
  },
```

## Related Issue

Fixes #1989 
<!-- or -->
Relates to #

## Type of change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [x] Other (security config, docs update, etc)

## Checklist before merging
- [x] Unit,
[Journey](https://github.com/defenseunicorns/pepr/tree/main/journey),
[E2E Tests](https://github.com/defenseunicorns/pepr-excellent-examples),
[docs](https://github.com/defenseunicorns/pepr/tree/main/docs),
[adr](https://github.com/defenseunicorns/pepr/tree/main/adr) added or
updated as needed
- [x] [Contributor Guide
Steps](https://docs.pepr.dev/main/contribute/#submitting-a-pull-request)
followed

---------

Signed-off-by: Case Wylie <cmwylie19@defenseunicorns.com>
Co-authored-by: Sam Mayer <sam.mayer@defenseunicorns.com>