chore: remove unnecessary prettier ignore rule (#2133)
## Description

This PR removes an unnecesary ignore rule. `data.json` is parsed by
scripts when used and is an ignored file under version control. In the
event a human has to examine the file when troubleshooting, a pretty
version of the file would prove useful.

## Related Issue

Relates to #1911

## Type of change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [x] Other (security config, docs update, etc)

## Checklist before merging
- [x] Unit,
[Journey](https://github.com/defenseunicorns/pepr/tree/main/journey),
[E2E Tests](https://github.com/defenseunicorns/pepr-excellent-examples),
[docs](https://github.com/defenseunicorns/pepr/tree/main/docs),
[adr](https://github.com/defenseunicorns/pepr/tree/main/adr) added or
updated as needed
- [x] [Contributor Guide
Steps](https://docs.pepr.dev/main/contribute/#submitting-a-pull-request)
followed