chore: in faq give explanation of module author (#2201)
## Description

We have been reaching out to the [Pepr
Community](https://kubernetes.slack.com/archives/C06DGH40UCB/p1747169572692789?thread_ts=1744212295.406759&cid=C06DGH40UCB)
in slack for feedback on our docs. One piece of feedback was that it is
not clear what a module author is. This PR adds to the faq an
explanation of a module author

## Related Issue

Fixes #2172 
<!-- or -->
Relates to #

## Type of change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [x] Other (security config, docs update, etc)

## Checklist before merging
- [x] Unit,
[Journey](https://github.com/defenseunicorns/pepr/tree/main/journey),
[E2E Tests](https://github.com/defenseunicorns/pepr-excellent-examples),
[docs](https://github.com/defenseunicorns/pepr/tree/main/docs),
[adr](https://github.com/defenseunicorns/pepr/tree/main/adr) added or
updated as needed
- [x] [Contributor Guide
Steps](https://docs.pepr.dev/main/contribute/#submitting-a-pull-request)
followed

---------

Signed-off-by: Case Wylie <cmwylie19@defenseunicorns.com>
Co-authored-by: Sam Mayer <sam.mayer@defenseunicorns.com>