chore: do not log about an alias when there is no alias (#2203)
## Description

When debugging a Pepr module I noticed this useless log polluting the
controller logs repeatedly.

```bash
Executing reconcile action with alias: no alias provided
```

For one, it is incorrect, there is no alias in the action, two it is
pointless, three this is a `Log.info` instead of debug, that kind of
verbosity is bad for debugging and user experience as it does not
provide any useful information.

If there is an alias. we can use this and it could be useful for
grepping logs but without an alias it just does not make sense

## Related Issue

Fixes #2202
<!-- or -->
Relates to #

## Type of change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [x] Other (security config, docs update, etc)

## Checklist before merging
- [x] Unit,
[Journey](https://github.com/defenseunicorns/pepr/tree/main/journey),
[E2E Tests](https://github.com/defenseunicorns/pepr-excellent-examples),
[docs](https://github.com/defenseunicorns/pepr/tree/main/docs),
[adr](https://github.com/defenseunicorns/pepr/tree/main/adr) added or
updated as needed
- [x] [Contributor Guide
Steps](https://docs.pepr.dev/main/contribute/#submitting-a-pull-request)
followed

---------

Signed-off-by: Case Wylie <cmwylie19@defenseunicorns.com>