chore: refactor tests config complexity in preparation for kfc latest and v1 (#2278)
## Description

This PR includes a refactor of unit tests, configurations for unit,
integration, journey, TypeScript, and ESLint in order to bring in the
latest version of kubernetes-fluent-client which is now an ESM module
and will not work in Pepr.

A separate `tsconfig.json` is necessary because with a single one, the
build fails:

```bash
src/sdk/sdk.test.ts:122:5 - error TS2739: Type 'Mock<Procedure>' is missing the following properties from type 'Mock<UnknownFunction>': _isMockFunction, _protoImpl

122     Create = vi.fn();
```

if you exclude the tests from `tsconfig.json` then you get errors about
the tests not being found by the project service:

```bash
src/sdk/sdk.test.ts was not found by the project service. Consider either including it in the tsconfig.json or including it in allowDefaultProject
```

So the best option was to create `tsconfig.test.json` specifically for
the tests.


## Related Issue

Fixes #2257 
<!-- or -->
Relates to #

## Type of change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Other (security config, docs update, etc)

## Checklist before merging
- [ ] Unit,
[Journey](https://github.com/defenseunicorns/pepr/tree/main/journey),
[E2E Tests](https://github.com/defenseunicorns/pepr-excellent-examples),
[docs](https://github.com/defenseunicorns/pepr/tree/main/docs),
[adr](https://github.com/defenseunicorns/pepr/tree/main/adr) added or
updated as needed
- [ ] [Contributor Guide
Steps](https://docs.pepr.dev/main/contribute/#submitting-a-pull-request)
followed

---------

Signed-off-by: Case Wylie <cmwylie19@defenseunicorns.com>
Co-authored-by: Sam Mayer <sam.mayer@defenseunicorns.com>