feat: consider whether bindings are watch or admission during manifests generation (#2359)
## Description

We already consider if a module is watch, and if the bindings are
admission then the watch manifests are not created. However, we deploy
the admission deployment and service regardless. Also, we are not taking
into account bindings during the helm chart generation. We need to
consider this when generating the `values.yaml` and default these not to
render based on the bindings.

PEXEX Tests
-
[e2e](https://github.com/defenseunicorns/pepr-excellent-examples/pull/338)
- [required fix for
moduleUp](https://github.com/defenseunicorns/pepr-excellent-examples/pull/337)

## Related Issue

Fixes #2349 
Fixes #2340 

<!-- or -->
Relates to #

## Type of change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Other (security config, docs update, etc)

## Checklist before merging
- [x] Unit,
[Journey](https://github.com/defenseunicorns/pepr/tree/main/journey),
[E2E Tests](https://github.com/defenseunicorns/pepr-excellent-examples),
[docs](https://github.com/defenseunicorns/pepr/tree/main/docs),
[adr](https://github.com/defenseunicorns/pepr/tree/main/adr) added or
updated as needed
- [x] [Contributor Guide
Steps](https://docs.pepr.dev/main/contribute/#submitting-a-pull-request)
followed

---------

Signed-off-by: Case Wylie <cmwylie19@defenseunicorns.com>