chore: remove TEST_MODE environment variable (#2449)
## Description

This PR is part of #2281 and removes an environment-variable based
approach to testing. Instead, we test the Pepr CLI the same way a user
would.

Of note, we remove the `TEST_MODE` envar (finally) and test the CLI as
if we were an actual user in `/journey`` until we can finally remove
that directory.

## Related Issue

Relates to #2281 

## Type of change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [x] Other (security config, docs update, etc)

## Checklist before merging
- [x] Unit,
[Journey](https://github.com/defenseunicorns/pepr/tree/main/journey),
[E2E Tests](https://github.com/defenseunicorns/pepr-excellent-examples),
[docs](https://github.com/defenseunicorns/pepr/tree/main/docs),
[adr](https://github.com/defenseunicorns/pepr/tree/main/adr) added or
updated as needed
- [x] [Contributor Guide
Steps](https://docs.pepr.dev/main/contribute/#submitting-a-pull-request)
followed

---------

Co-authored-by: Case Wylie <cmwylie19@defenseunicorns.com>