chore: lower warn threshold for changed files after release (#2576)
## Description

This PR lowers the warn threshold back to 15 following a release of Pepr
that included changes to lots of files.

We use this warning to prevent us from unintentionally including
irrelevant files in the build and to encourage us to keep the scope of
work small by default.

## Related Issue

None

## Type of change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [x] Other (security config, docs update, etc)

## Checklist before merging
- [x] Unit, Integration,
[docs](https://github.com/defenseunicorns/pepr/tree/main/docs),
[adr](https://github.com/defenseunicorns/pepr/tree/main/adr) added or
updated as needed
- [x] [Contributor Guide
Steps](https://docs.pepr.dev/main/contribute/#submitting-a-pull-request)
followed