chore: teardowntimeout and workers for tutorials test (#2687)
## Description

The tutorials test has been failing in CI and then I noticed it also
failed locally. After adding this change it has passed 3 times.

This seems to make it more stable. 

Once this PR proves to make the test more stable then we should ensure
the release is blocked if this test fails. The funny thing is now 24/24
tests are passing but the error comes from an `onTaskUpdate` timeout.


## Related Issue

Fixes #2685 
<!-- or -->
Relates to #2686 

## Type of change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [x] Other (security config, docs update, etc)

## Checklist before merging
- [x] Unit, Integration,
[docs](https://github.com/defenseunicorns/pepr/tree/main/docs),
[adr](https://github.com/defenseunicorns/pepr/tree/main/adr) added or
updated as needed
- [x] [Contributor Guide
Steps](https://docs.pepr.dev/main/contribute/#submitting-a-pull-request)
followed

Signed-off-by: Case Wylie <cmwylie19@defenseunicorns.com>