fix: run semantic release action from merge queue (#2697)
## Description

We need to run the semantic release from a merge queue so that it can
create the GitHub release, trigger the current release workflow. WIP

[Original
PR](https://github.com/defenseunicorns/pepr/actions/runs/18572172001/job/52948539310)
- Our release scripts and actions are expecting release events, and on
push to main the refs are different and causes problems.

Proposed solution - Keep the release action the exact same as it was,
and have semantic release occur in the merge queue, kicking off a GitHub
release, and the original release action works like normal

**Background**
When a PR goes into the merge queue, the code is updated and CI runs on
the updated code on the branch. We are running unit tests, integration
tests, several flavors of PEXEX tests ( unicorn, iron bank, upstream ).
This is overkill as all of these tests would have _just passed_ before
entering into the merge queue. The _risk_ we have to account for is
really in the _build_. If the PR that goes into a queue is a TypeScript
update and the next PR in the queue is a unit test, then once the unit
test PR is updated with the TypeScript update there could be _version
conflicts_. So what we really need to do is test that the code can build
and all of the unit and integration tests pass. Not retest the
controller.

## Related Issue

Fixes #
<!-- or -->
Relates to #

## Type of change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Other (security config, docs update, etc)

## Checklist before merging
- [ ] Unit, Integration,
[docs](https://github.com/defenseunicorns/pepr/tree/main/docs),
[adr](https://github.com/defenseunicorns/pepr/tree/main/adr) added or
updated as needed
- [ ] [Contributor Guide
Steps](https://docs.pepr.dev/main/contribute/#submitting-a-pull-request)
followed

---------

Signed-off-by: Case Wylie <cmwylie19@defenseunicorns.com>